+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- `Browser.get` and `Tab.close` will now wait for their appropiate target events before returning @ccev
+
 ### Added
 
 ### Changed